src/hispec/util/xeryon/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xeryon"
version = "0.1.0"
dependencies = [
  "pyserial"
]
description = "This module provides a Python interface to communicate with and control Xeryon precision stages. It supports serial communication, axis movement, settings management, and safe handling of errors and edge cases."
authors = [
    { name = "Michael Langmayr", email = "langmayr@caltech.edu" }
]
maintainers = [
    {name = "Michael Langmayr", email = "langmayr@caltech.edu"}
]
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
classifiers = [
  "Programming Language :: Python"
]

[project.urls]
# Various URLs related to your project. These links are displayed on PyPI.
# Homepage = "https://example.com"
# Documentation = "https://readthedocs.org"
Repository = "https://github.com/COO-Utils/xeryon"
# "Bug Tracker" = "https://github.com/COO-Utils/xeryon/issues"
# Changelog = "https://github.com/yourusername/your-repo/blob/master/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest-mock",
    "pytest", 
    "black",
    "flake8"
]
[tool.setuptools.packages.find]
include = ["axis", "config", "units", "utils", "communication", "stage", "xeryon_controller"]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
py-modules = ["xeryon"]
